\name{nearfar-package}
\alias{nearfar-package}
\alias{nearfar}
\docType{package}
\title{
\packageTitle{nearfar}
}
\description{
\packageDescription{nearfar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nearfar}
\packageIndices{nearfar}

Please see \code{\link{opt.nearfar}} for a detailed example displaying
full functionality of the nearfar package.

The authors graciously acknowledge support for this work from the
National Institute On Minority Health And Health Disparities of the
National Institutes of Health under Award Number DP2MD010478 and the
Agency for Healthcare Research and Quality of the National Institutes of
Health under Award Number 5R00HS022192-03.

}
\author{
\packageAuthor{nearfar}

Maintainer: \packageMaintainer{nearfar}
}
\references{

Rigdon J, Baiocchi M, Basu S (2016). Near-far matching in R: The nearfar
package.  Submitted to the Journal of Statistical Software.

Baiocchi M, Small D, Lorch S, Rosenbaum P (2010). Building a stronger instrument in
an observational study of perinatal care for premature infants. Journal of the American
Statistical Association, 105(492), 1285-1296.

Baiocchi M, Small D, Yang L, Polsky D, Groeneveld P (2012). Near-far matching: a study design approach to instrumental variables. Health Services and Outcomes Research Methodology, 12(4), 237-253.

}
