\name{nearfar-package}
\alias{nearfar-package}
\docType{package}
\title{Near-Far Matching}
\description{Near-far matching is a study design technique for
    preprocessing observational data to mimic a pair-randomized trial.
    Individuals are matched to be near on measured confounders and far
    on levels of an instrumental variable.}

\details{
\tabular{ll}{
Package: \tab nearfar\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2017-06-01\cr
License: \tab GPL-3\cr
}
}

\author{Joseph Rigdon \email{jrigdon@stanford.edu}}
\references{

Rigdon J, Baiocchi M, Basu S (2018). Near-far matching in R: The nearfar
package. Journal of Statistical Software, 86(5), 1-21.

Baiocchi M, Small D, Lorch S, Rosenbaum P (2010). Building a
stronger instrument in an observational study of perinatal care for
premature infants. Journal of the American Statistical Association,
105(492), 1285-1296.

Baiocchi M, Small D, Yang L, Polsky D, Groeneveld P (2012). Near-far
matching: a study design approach to instrumental variables. Health
Services and Outcomes Research Methodology, 12(4), 237-253.

}
